third_party/blink/web_tests/screen_orientation/screenorientation-unsupported-no-crash.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/testdriver.js"></script>
<script src="../resources/testdriver-vendor.js"></script>
<script src="../fast/forms/resources/picker-common.js"></script>
<input type=date id=date>
<script>
test(function() {
    function popupCallback() {
        assert_equals(popupWindow.screen.orientation, null);
    }
    openPicker(document.getElementById('date'), popupCallback);
}, "Verify that accessing screen.orientation where unsupported doesn't crash.");
</script>
</body>
</html>
